% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiasocgenes.R
\name{epiasocgenes}
\alias{epiasocgenes}
\title{epiasocgenes}
\usage{
epiasocgenes(bfile, out, primarylist = NULL, genelist = NULL,
  plink = "plink", core = detectCores() - 1,
  varpl = "annotate_variation.pl", ANNOVAR = T, clump = T,
  build = "hg19",
  db = "../../Doctorado/Annovar/annovar.latest/annovar/humandb/",
  Wu = T, covar = NULL, ncov = "1-2", AsT = 1e-04, MAF = 0.01)
}
\arguments{
\item{bfile}{Es la ruta del archivo bfile de Plink.}

\item{out}{Es la ruta al directorio donde se guardará la salida de la función.}

\item{primarylist}{Es la rutal al archivo con los genes en los que deseamos que se realice la búsqueda de SNPs primarios.}

\item{genelist}{Es la ruta al archivo con los genes en los que deseamos que se realice la búsqueda de epistasis o SNPs secundarios.}

\item{plink}{Es la ruta al ejecutable de plink.}

\item{core}{Por defecto selecciona todos los cores menos uno.}

\item{varpl}{Es la ruta al archivo annotate_variation.pl de ANNOVAR.}

\item{ANNOVAR}{Lógico, se recomienda que sea T. Anotación de las variantes con ANNOVAR.}

\item{clump}{Lógico, se recomienda que sea T si existen más de 2000 sujetos, realiza un clumping.}

\item{build}{Es la versión del genoma de referencia.}

\item{db}{Es la ruta al directorio donde se encuentra la base de datos de ANNOVAR.}

\item{Wu}{Lógico, se recomienda que sea T. Realiza la corrección de Wu en epistasis.}

\item{covar}{Es la ruta al archivo de covariables.}

\item{ncov}{Es el índice de las covariables a introducir en el análisis.}

\item{AsT}{Por defecto 0.0001. Es el límite de p.valor en el que se seleccionan SNPs primarios.}

\item{MAF}{Por defecto 0.01. Es el limite de MAF en el que se seleccionan SNPs primarios.}
}
\description{
Esta función realiza un test de epistasis seleccionando SNPs prometedores.
}
\details{
Esta funcion es parte del doctorado de A.Cisterna para realizar un test de epistasis.
La documentacion de ayuda es generada usando roxygen2.
}
